requirements.txt
streamlit
pandas
psycopg2-binary
scipy
requests
matplotlib
pytz
